# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mnemos API",
    "description": "Knowledge management and RAG (Retrieval-Augmented Generation) system with multi-modal chat, document processing, knowledge graphs, and video mixing capabilities.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Backend server"
    },
    {
      "url": "http://localhost:5200",
      "description": "Frontend proxy"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication (placeholder for future implementation)"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key authentication (placeholder for future implementation)"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Resource not found"
          }
        }
      }
    }
  },
  "paths": {
    "/api/documents/": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "List all documents",
        "description": "Retrieve all documents with optional filtering by collection. Returns HTML partial for HTMX requests or JSON.",
        "parameters": [
          {
            "name": "collection_id",
            "in": "query",
            "description": "Filter by collection ID (use 'null' for uncollected documents)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/search": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Search passages (LLM-free)",
        "description": "Hybrid vector + keyword retrieval that returns ranked passages without any LLM generation. Works even when no language model is configured. Scope with document_ids, or omit to search the whole library.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "query"
                ],
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "Search query",
                    "example": "snafu principle"
                  },
                  "document_ids": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "description": "Restrict search to these document IDs (optional)"
                  },
                  "top_k": {
                    "type": "integer",
                    "default": 10,
                    "maximum": 50,
                    "description": "Maximum number of passages to return"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ranked passages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "query": {
                      "type": "string"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "content": {
                            "type": "string"
                          },
                          "chunk_index": {
                            "type": "integer",
                            "nullable": true
                          },
                          "page_number": {
                            "type": "integer",
                            "nullable": true
                          },
                          "start_time": {
                            "type": "number",
                            "nullable": true
                          },
                          "end_time": {
                            "type": "number",
                            "nullable": true
                          },
                          "document_id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "document_title": {
                            "type": "string",
                            "nullable": true
                          },
                          "file_type": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/upload": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Upload a document",
        "description": "Upload a file or provide a YouTube URL for processing. Supported formats: PDF, EPUB, audio (MP3, WAV, M4A), video (MP4, WebM, MOV).",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "File to upload"
                  },
                  "youtube_url": {
                    "type": "string",
                    "description": "YouTube URL (alternative to file upload)",
                    "example": "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Document created and queued for processing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "400": {
            "description": "No file or URL provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{doc_id}": {
      "put": {
        "tags": [
          "Documents"
        ],
        "summary": "Update document metadata",
        "description": "Update document tags, stars, comments, or collection assignment.",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tag": {
                    "type": "string",
                    "example": "important"
                  },
                  "stars": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 5,
                    "example": 5
                  },
                  "comment": {
                    "type": "string",
                    "example": "Key reference material"
                  },
                  "collection_id": {
                    "type": "string",
                    "format": "uuid",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "404": {
            "description": "Document not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Documents"
        ],
        "summary": "Delete a document",
        "description": "Delete document and associated files from the system.",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document deleted successfully"
          },
          "404": {
            "description": "Document not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/documents/{doc_id}/status": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Get document processing status",
        "description": "Get current processing status (for HTMX polling).",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{doc_id}/content": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "Download document file",
        "description": "Download the original document file.",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File download",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Document not found"
          }
        }
      }
    },
...
  }
}